Added product info

+++ src/main/java/com/dailycodebuffer/springbootdemo/repository/jpa/CourseMaterialEntityRepository.java
@@ -3,5 +3,5 @@ package com.dailycodebuffer.springbootdemo.repository.jpa;
 import com.dailycodebuffer.springbootdemo.entities.jpa.CourseMaterialEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseMaterialEntityRepository extends JpaRepository<CourseMaterialEntity, Long> {
+public interface CourseMaterialEntityRepository extends JpaRepository<CourseMaterialEntity, Integer> {
 }
